+++ source/renderer/app/components/staking/stake-pools/StakePoolsSearch.js
@@ -7,6 +7,8 @@ import { InputSkin } from 'react-polymorph/lib/skins/simple/InputSkin';
 import { PopOver } from 'react-polymorph/lib/components/PopOver';
 import classnames from 'classnames';
 import styles from './StakePoolsSearch.scss';
+import FilterPopOver from './FilterPopOver';
+import type { FilterPopOverProps } from './FilterPopOver';
 import searchIcon from '../../../assets/images/search.inline.svg';
 import closeIcon from '../../../assets/images/close-cross.inline.svg';
 import gridIcon from '../../../assets/images/grid-ic.inline.svg';
@@ -52,6 +54,9 @@ const messages = defineMessages({
   },
 });
 
+const INPUT_FIELD_PADDING_DELTA = 60;
+const CLEAR_BUTTON_RIGHT_DELTA = 10;
+
 type Props = {
   label?: string,
   placeholder?: string,
@@ -64,6 +69,7 @@ type Props = {
   onGridRewardsView?: Function,
   onListView?: Function,
   search: string,
+  filterPopOverProps: FilterPopOverProps,
 };
 
 export class StakePoolsSearch extends Component<Props> {
@@ -72,6 +78,15 @@ export class StakePoolsSearch extends Component<Props> {
   };
 
   searchInput: ?Object = null;
+  addOnRef: { current: null | HTMLDivElement };
+  selfRef: { current: null | HTMLDivElement };
+
+  constructor(props: Props) {
+    super(props);
+
+    this.addOnRef = React.createRef();
+    this.selfRef = React.createRef();
+  }
 
   autoSelectOnFocus = () =>
     this.searchInput ? this.searchInput.inputElement.current.select() : false;
@@ -87,6 +102,23 @@ export class StakePoolsSearch extends Component<Props> {
     }
   };
 
+  generateDynamicStyles = () => {
+    const { current: addOnDom } = this.addOnRef;
+    if (!addOnDom) {
+      return null;
+    }
+
+    const addOnDomRect = addOnDom.getBoundingClientRect();
+    return {
+      inputField: {
+        paddingRight: `${addOnDomRect.width + INPUT_FIELD_PADDING_DELTA}px`,
+      },
+      clearButton: {
+        right: `${addOnDomRect.width + CLEAR_BUTTON_RIGHT_DELTA}px`,
+      },
+    };
+  };
+
   render() {
     const { intl } = this.context;
     const {
@@ -100,7 +132,9 @@ export class StakePoolsSearch extends Component<Props> {
       isListView,
       isGridView,
       isGridRewardsView,
+      filterPopOverProps,
     } = this.props;
+    const dynamicStyles = this.generateDynamicStyles();
 
     const gridButtonClasses = classnames([
       styles.gridView,
@@ -119,26 +153,14 @@ export class StakePoolsSearch extends Component<Props> {
 
     const isBigSearchComponent = isListView || isGridView || isGridRewardsView;
 
-    const searchInputClases = classnames([
-      styles.searchInput,
-      isBigSearchComponent ? styles.inputExtrasSearch : null,
-      IS_GRID_REWARDS_VIEW_AVAILABLE ? styles.withGridRewardsView : null,
-    ]);
-
-    const clearSearchClasses = classnames([
-      styles.inputExtras,
-      isBigSearchComponent ? styles.inputExtrasSearch : null,
-      IS_GRID_REWARDS_VIEW_AVAILABLE ? styles.withGridRewardsView : null,
-    ]);
-
     return (
-      <div className={styles.component}>
+      <div className={styles.component} ref={this.selfRef}>
         <div className={styles.container}>
           <SVGInline svg={searchIcon} className={styles.searchIcon} />
           <Input
             autoFocus
             label={label || null}
-            className={searchInputClases}
+            className={styles.searchInput}
             onChange={onSearch}
             ref={(input) => {
               this.searchInput = input;
@@ -150,9 +172,13 @@ export class StakePoolsSearch extends Component<Props> {
             value={search}
             maxLength={150}
             onFocus={this.autoSelectOnFocus}
+            style={dynamicStyles ? dynamicStyles.inputField : null}
           />
           {this.hasSearchClearButton && (
-            <div className={clearSearchClasses}>
+            <div
+              className={styles.inputExtras}
+              style={dynamicStyles ? dynamicStyles.clearButton : null}
+            >
               <PopOver content={intl.formatMessage(messages.clearTooltip)}>
                 <button
                   onClick={this.handleClearSearch}
@@ -167,7 +193,7 @@ export class StakePoolsSearch extends Component<Props> {
             </div>
           )}
           {isBigSearchComponent && (
-            <div className={styles.viewButtons}>
+            <div className={styles.viewButtons} ref={this.addOnRef}>
               <span className={styles.separator}>|</span>
               <PopOver content={intl.formatMessage(messages.gridIconTooltip)}>
                 <button className={gridButtonClasses} onClick={onGridView}>
@@ -191,6 +217,11 @@ export class StakePoolsSearch extends Component<Props> {
                   <SVGInline svg={listIcon} />
                 </button>
               </PopOver>
+              <span className={styles.separator}>|</span>
+              <FilterPopOver
+                {...filterPopOverProps}
+                containerRefDom={this.selfRef.current}
+              />
             </div>
           )}
         </div>