+++ source/renderer/app/components/staking/widgets/ThumbSelectedPool.js
@@ -5,6 +5,7 @@ import SVGInline from 'react-svg-inline';
 import classnames from 'classnames';
 import styles from './ThumbSelectedPool.scss';
 import { getColorFromRange } from '../../../utils/colors';
+import { ellipsis } from '../../../utils/strings';
 import checkmarkImage from '../../../assets/images/check-w.inline.svg';
 import questionmarkImage from '../../../assets/images/questionmark.inline.svg';
 import clockImage from '../../../assets/images/clock.inline.svg';
@@ -13,6 +14,7 @@ import { IS_RANKING_DATA_AVAILABLE } from '../../../config/stakingConfig';
 
 type Props = {
   stakePool?: StakePool,
+  ranking: number,
   alreadyDelegated?: boolean,
   numberOfRankedStakePools: number,
 };
@@ -22,11 +24,12 @@ export default class ThumbSelectedPool extends Component<Props> {
   render() {
     const {
       stakePool,
+      ranking,
       alreadyDelegated,
       numberOfRankedStakePools,
     } = this.props;
 
-    const { ticker, retiring, ranking } = stakePool || {};
+    const { id, ticker, retiring } = stakePool || {};
     const rankColor =
       stakePool && !retiring && IS_RANKING_DATA_AVAILABLE
         ? getColorFromRange(ranking, numberOfRankedStakePools)
@@ -54,7 +57,11 @@ export default class ThumbSelectedPool extends Component<Props> {
 
     return (
       <div className={selectedPoolBlockClasses} style={selectedPoolBlockStyle}>
-        {ticker && <div className={styles.ticker}>{ticker}</div>}
+        {ticker ? (
+          <div className={styles.ticker}>{ticker}</div>
+        ) : (
+          <div className={styles.id}>{ellipsis(id, 4, 4)}</div>
+        )}
         <div className={styles.icon}>
           <SVGInline svg={icon} />
         </div>