+++ source/renderer/app/containers/staking/dialogs/DelegationSetupWizardDialogContainer.js
@@ -176,6 +176,7 @@ export default class DelegationSetupWizardDialogContainer extends Component<
       networkStatus,
       hardwareWallets,
     } = this.props.stores;
+    const { staking: stakingActions } = this.props.actions;
     const { futureEpoch } = networkStatus;
     const { currentTheme, currentLocale } = profile;
     const {
@@ -190,7 +191,10 @@ export default class DelegationSetupWizardDialogContainer extends Component<
       joinStakePoolRequest,
       getStakePoolById,
       isDelegationTransactionPending,
+      filterOptions,
+      populatedFilterOptions,
     } = staking;
+    const { filterStakePools } = stakingActions;
     const futureEpochStartTime = get(futureEpoch, 'epochStart', 0);
     const selectedPool = find(stakePools, (pool) => pool.id === selectedPoolId);
 
@@ -244,6 +248,9 @@ export default class DelegationSetupWizardDialogContainer extends Component<
           selectCoinsRequest.error
         }
         hwDeviceStatus={hwDeviceStatus}
+        filterOptions={filterOptions || {}}
+        populatedFilterOptions={populatedFilterOptions}
+        onFilter={filterStakePools.trigger}
         isTrezor={isTrezor}
       />
     );