+++ storybook/stories/staking/DelegationSteps.stories.js
@@ -173,6 +173,9 @@ export class StakingDelegationSteps extends Component<Props, State> {
         onSelectPool={this.onContinue}
         selectedPool={null}
         selectedWallet={WALLETS[0]}
+        filterOptions={{}}
+        onFilter={() => null}
+        populatedFilterOptions={{}}
       />,
       <DelegationStepsConfirmationDialog
         key="DelegationStepsConfirmationDialog"